% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.r2cdocs-package.R
\docType{package}
\name{igraph.r2cdocs-package}
\alias{igraph.r2cdocs}
\alias{igraph.r2cdocs-package}
\title{igraph.r2cdocs: Link C docs for the igraph package}
\description{
Holds infrastructure for a special roclet that lets us link to C docs from R manual pages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/igraph/igraph.r2cdocs}
  \item Report bugs at \url{https://github.com/igraph/igraph.r2cdocs/issues}
}

}
\author{
\strong{Maintainer}: Maëlle Salmon \email{msmaellesalmon@gmail.com} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})

}
\keyword{internal}
